Rejig extensions layout
